Incrementando Dialog no processo de exclusão de registro.

+++ lib/screens/home_screen/widgets/journal_card.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_webapi_first_course/helpers/weekday.dart';
 import 'package:flutter_webapi_first_course/models/journal.dart';
+import 'package:flutter_webapi_first_course/screens/commom/confirmate_dialog.dart';
 import 'package:flutter_webapi_first_course/services/journal_service.dart';
 import 'package:uuid/uuid.dart';
 
@@ -142,13 +143,24 @@ class JournalCard extends StatelessWidget {
 
   removeJournal(BuildContext context) {
     JournalService service = JournalService();
-    service.delete(journal!.id).then((value) {
-      if (value) {
-        ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(content: Text("Removido com sucesso!")));
 
-        refreshFunction();
-      }
-    });
+    showConfirmationDialog(
+      context,
+      content:
+          "Deseja realmente remover o diário do dia ${WeekDay(journal!.createdAt)}?",
+      affirmativeOption: "Remover",
+    ).then((value) => {
+          if (value != null && value)
+            {
+              service.delete(journal!.id).then((value) {
+                if (value) {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                      const SnackBar(content: Text("Removido com sucesso!")));
+
+                  refreshFunction();
+                }
+              })
+            }
+        });
   }
 }
